# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "JWT Authorization header using the Bearer scheme. Example: \"Authorization",
        "title": "RAVR Backend API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "url": "http://www.ravr-site.io/support",
            "email": "support@ravr-site.io"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/api/v1",
    "paths": {
...
        "/auth/login": {
            "post": {
                "description": "Аутентифицирует пользователя и возвращает JWT токен",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "auth"
                ],
                "summary": "Вход в систему",
                "parameters": [
                    {
                        "description": "Учетные данные",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/controller.loginRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/controller.SwaggerTokenResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/responses.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/responses.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/auth/register": {
            "post": {
                "description": "Создает нового пользователя с указанными учетными данными",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "auth"
                ],
                "summary": "Регистрация нового пользователя",
                "parameters": [
                    {
                        "description": "Данные для регистрации",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/controller.registerRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/controller.SwaggerMessageResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/responses.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/auth/user": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Возвращает данные текущего аутентифицированного пользователя с статистикой",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "auth"
                ],
                "summary": "Профиль пользователя",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/controller.SwaggerUserProfileWithStatsResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/responses.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/responses.ErrorResponse"
                        }
                    }
                }
            },
            "put": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Обновляет имя и фамилию текущего аутентифицированного пользователя",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "auth"
                ],
                "summary": "Обновление данных пользователя",
                "parameters": [
                    {
                        "description": "Данные для обновления",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/controller.updateUserRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/controller.SwaggerUpdateUserResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/responses.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/responses.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/responses.ErrorResponse"
                        }
                    }
                }
            }
        },
...
        "/lessons/complete": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Обновляет статистику пользователя при завершении урока и возвращает лидерборд по уроку",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "lessons"
                ],
                "summary": "Завершение урока",
                "parameters": [
                    {
                        "description": "Данные о прохождении урока",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/controller.CompleteRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/controller.CompleteLessonResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/responses.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/responses.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/responses.ErrorResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "controller.CompleteLessonResponse": {
            "type": "object",
            "properties": {
                "completedTime": {
                    "description": "Время завершения урока в секундах",
                    "type": "integer"
                },
                "earnedXP": {
                    "description": "Заработанный опыт за текущий урок",
                    "type": "integer"
                },
                "experience": {
                    "description": "Текущий опыт пользователя",
                    "type": "integer"
                },
                "leaderboard": {
                    "description": "Лидерборд по уроку",
                    "type": "object",
                    "properties": {
                        "entries": {
                            "description": "Записи лидерборда",
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/repository.LessonLeaderboardEntry"
                            }
                        },
                        "totalUsers": {
                            "description": "Общее количество пользователей",
                            "type": "integer"
                        },
                        "userPosition": {
                            "description": "Позиция пользователя",
                            "type": "integer"
                        }
                    }
                },
                "success": {
                    "type": "boolean"
                }
            }
        },
        "controller.CompleteRequest": {
            "type": "object",
            "properties": {
                "completionTime": {
                    "description": "Время в секундах",
                    "type": "integer"
                },
                "earnedExperience": {
                    "type": "integer"
                },
                "lessonId": {
                    "type": "string"
                }
            }
        },
...
        "controller.SwaggerMessageResponse": {
            "description": "Успешный ответ с сообщением",
            "type": "object",
            "properties": {
                "data": {
                    "type": "object",
                    "additionalProperties": {
                        "type": "string"
                    }
                },
                "success": {
                    "type": "boolean",
                    "example": true
                }
            }
        },
...
        "controller.SwaggerTokenResponse": {
            "description": "Успешный ответ с токеном авторизации",
            "type": "object",
            "properties": {
                "data": {
                    "type": "object",
                    "properties": {
                        "token": {
                            "type": "string",
                            "example": "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9..."
                        }
                    }
                },
                "success": {
                    "type": "boolean",
                    "example": true
                }
            }
        },
        "controller.SwaggerUpdateUserResponse": {
            "description": "Ответ с сообщением об успешном обновлении",
            "type": "object",
            "properties": {
                "data": {
                    "type": "object",
                    "properties": {
                        "message": {
                            "type": "string",
                            "example": "Данные пользователя успешно обновлены"
                        }
                    }
                },
                "success": {
                    "type": "boolean",
                    "example": true
                }
            }
        },
        "controller.SwaggerUserProfileWithStatsResponse": {
            "description": "Расширенный ответ с данными профиля пользователя и статистикой",
            "type": "object",
            "properties": {
                "data": {
                    "type": "object",
                    "properties": {
                        "experience": {
                            "type": "integer",
                            "example": 1250
                        },
                        "first_name": {
                            "type": "string",
                            "example": "John"
                        },
                        "id": {
                            "type": "integer",
                            "example": 1
                        },
                        "last_name": {
                            "type": "string",
                            "example": "Doe"
                        },
                        "stats": {
                            "type": "object",
                            "properties": {
                                "average_completion": {
                                    "type": "number",
                                    "example": 120.5
                                },
                                "average_experience": {
                                    "type": "number",
                                    "example": 83.33
                                },
                                "fastest_completion": {
                                    "type": "integer",
                                    "example": 85
                                },
                                "max_experience": {
                                    "type": "integer",
                                    "example": 150
                                },
                                "total_experience": {
                                    "type": "integer",
                                    "example": 1250
                                },
                                "total_lessons": {
                                    "type": "integer",
                                    "example": 15
                                }
                            }
                        },
                        "username": {
                            "type": "string",
                            "example": "johndoe"
                        }
                    }
                },
                "success": {
                    "type": "boolean",
                    "example": true
                }
            }
        },
        "controller.loginRequest": {
            "description": "Запрос на вход в систему",
            "type": "object",
            "required": [
                "password",
                "username"
            ],
            "properties": {
                "password": {
                    "type": "string",
                    "example": "secret123"
                },
                "username": {
                    "type": "string",
                    "example": "johndoe"
                }
            }
        },
        "controller.registerRequest": {
            "description": "Запрос на регистрацию нового пользователя",
            "type": "object",
            "required": [
                "password",
                "username"
            ],
            "properties": {
                "first_name": {
                    "type": "string",
                    "example": "John"
                },
                "last_name": {
                    "type": "string",
                    "example": "Doe"
                },
                "password": {
                    "type": "string",
                    "minLength": 6,
                    "example": "secret123"
                },
                "username": {
                    "type": "string",
                    "minLength": 3,
                    "example": "johndoe"
                }
            }
        },
        "controller.updateUserRequest": {
            "description": "Запрос на обновление данных пользователя",
            "type": "object",
            "properties": {
                "first_name": {
                    "type": "string"
                },
                "last_name": {
                    "type": "string"
                }
            }
        },
        "repository.LessonLeaderboardEntry": {
            "type": "object",
            "properties": {
                "completion_time": {
                    "description": "Время прохождения урока в секундах",
                    "type": "integer"
                },
                "experience": {
                    "description": "Полученный опыт",
                    "type": "integer"
                },
                "first_name": {
                    "type": "string"
                },
                "last_name": {
                    "type": "string"
                },
                "position": {
                    "type": "integer"
                },
                "score": {
                    "description": "Баллы за урок (время в секундах)",
                    "type": "integer"
                },
                "trend": {
                    "description": "Тренд изменения позиции",
                    "type": "string"
                },
                "user_id": {
                    "type": "integer"
                },
                "username": {
                    "type": "string"
                }
            }
        },
        "responses.ErrorInfo": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "string"
                },
                "message": {
                    "type": "string"
                }
            }
        },
        "responses.ErrorResponse": {
            "type": "object",
            "properties": {
                "error": {
                    "$ref": "#/definitions/responses.ErrorInfo"
                },
                "success": {
                    "type": "boolean"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "JWT Authorization header using the Bearer scheme. Example: \"Authorization: Bearer \u003ctoken\u003e\"",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}